urca/urcacron/requirements.txt
absl-py==2.1.0
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
anyio==4.6.0
APScheduler==3.9.1.post1
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
Babel==2.11.0
backtrader==1.9.78.123
bcrypt==4.0.1
beautifulsoup4==4.12.3
bidict==0.23.1
black==24.4.2
blinker==1.8.1
botocore==1.34.84
bs4==0.0.2
cachetools==5.3.3
ccxt==4.3.41
certifi==2022.9.24
cffi==1.15.1
cfgv==3.4.0
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.2.1
croniter==2.0.5
cryptography==38.0.4
cycler==0.12.1
dateutils==0.6.12
distlib==0.3.8
dnspython==2.6.1
docx==0.2.4
docxcompose==1.3.7
docxtpl==0.16.4
endesive==2.0.12
espaco_exclusivo_package==0.1.32
et-xmlfile==1.1.0
eventlet==0.36.1
exceptiongroup==1.2.2
fastquant==0.1.8.1
filelock==3.14.0
Flask==2.2.2
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.51.0
fpdf==1.7.2
frozendict==2.4.4
frozenlist==1.4.1
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.64.1
gunicorn==20.1.0
h11==0.14.0
h5py==3.11.0
hmmlearn==0.3.2
html5lib==1.1
httpcore==0.16.3
httpx==0.23.3
identify==2.5.36
idna==3.4
importlib-metadata==5.1.0
importlib_resources==6.4.0
itsdangerous==2.1.2
jax==0.4.29
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kaleido==0.2.1
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
lightgbm==4.3.0
lxml==4.9.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.4.0
mpld3==0.5.10
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
nltk==3.8.1
nodeenv==1.9.1
numpy==1.23.5
oauthlib==3.2.2
openpyxl==3.1.0
opt-einsum==3.3.0
optree==0.11.0
oscrypto==1.3.0
packaging==24.0
pandas==1.5.3
paramiko==2.12.0
pathspec==0.12.1
peewee==3.17.5
Pillow==9.3.0
platformdirs==4.2.2
plotly==5.22.0
pre-commit==3.7.1
protobuf==4.25.3
psutil==5.9.8
pyarrow==16.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycares==4.4.0
pycparser==2.21
pycryptodome==3.16.0
pydeck==0.9.0
Pygments==2.17.2
PyGreSQL==6.0
PyKCS11==1.5.11
pymssql==2.3.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-bcb==0.1.8
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
reportlab==4.0.9
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.6.0
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
streamlit==1.34.0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.4
tweepy==4.14.0
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==4.2
Unidecode==1.3.6
urllib3==1.26.13
virtualenv==20.26.2
webencodings==0.5.1
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
xgboost==2.0.3
xlrd==2.0.1
XlsxWriter==3.0.3
yarl==1.9.4
yfinance==0.2.40
zipp==3.11.0
